+++ src/pages/Login.tsx
@@ -1,20 +1,14 @@
 import LoginFormContainer from '@components/Login/LoginFormContainer';
 import NavigationBar from '@components/Common/NavigationBar';
-import { useNavigate } from 'react-router-dom';
-import { URLS } from '@assets/urls';
 import StyledTopestDiv from '@src/styles/StyledTopesDiv';
 import { StyledLoginComponentContainerDiv } from '@src/components/Login/StyledLoginFormContainer';
 
-const Login: React.FC = () => {
-  const navigate = useNavigate();
-  const handleLoginSuccess = () => {
-    navigate(URLS.home);
-  };
+const Login = () => {
   return (
     <StyledTopestDiv>
       <NavigationBar />
       <StyledLoginComponentContainerDiv>
-        <LoginFormContainer onLoginSuccess={handleLoginSuccess} />
+        <LoginFormContainer />
       </StyledLoginComponentContainerDiv>
     </StyledTopestDiv>
   );
